Changed solutions to always return 64bit results and removed Result structs.

+++ project-euler.d/source/solutions/_11.d
@@ -13,7 +13,7 @@ mixin prepare_data!(11,
 
 mixin solution!(11,
 {
-	int max;
+	long max;
 	
 	for (int y = 0; y < 20; y++) {
 		for (int x = 0; x < 17; x++) {
